-Waves are no longer hard-coded (they are stored as .lvl files). I have them in the Debug folder, in a 'Game Levels' folder, which then has a folder with a specific 'level' (ex: Level0), which has 1 or more waves (ex: wave0.lvl + wave1.lvl) -Also, the gameObject class's X and Y are floats to allow slower/ more precise movement than we have been normally allowed. I have updated most speed atributes to be floats now. ~Torey

+++ CakeDefense/CakeDefense/CakeDefense/Bullet.cs
@@ -22,7 +22,7 @@ namespace CakeDefense
         #endregion Attributes
 
         #region Constructor
-        public Bullet(int dir, int dmg, int sp, int x, int y, int w, int h, Texture2D texture, SpriteBatch sprite)
+        public Bullet(int dir, int dmg, float sp, int x, int y, int w, int h, Texture2D texture, SpriteBatch sprite)
             :base(new ImageObject(texture, x, y, w, h, sprite), 0, dmg, sp)
         {
             IsActive = true;
